+++ android/src/main/kotlin/hamza/dali/flutter_osm_plugin/FlutterMaker.kt
@@ -4,12 +4,10 @@ import android.app.Application
 import android.content.Context
 import android.graphics.Bitmap
 import android.graphics.Canvas
-import android.graphics.Paint
-import android.graphics.Point
+import android.graphics.Matrix
 import android.graphics.drawable.BitmapDrawable
 import android.graphics.drawable.Drawable
 import android.view.LayoutInflater
-import android.view.MotionEvent
 import android.view.View
 import androidx.core.content.ContextCompat
 import androidx.core.graphics.BlendModeColorFilterCompat
@@ -21,7 +19,8 @@ import org.osmdroid.views.overlay.infowindow.MarkerInfoWindow
 
 typealias LongClickHandler = (marker: Marker) -> Boolean
 
-open class FlutterMarker(mapView: MapView) : Marker(mapView), Marker.OnMarkerClickListener, (Marker) -> Boolean {
+open class FlutterMarker(mapView: MapView) : Marker(mapView), Marker.OnMarkerClickListener,
+        (Marker) -> Boolean {
     protected lateinit var application: Application
     private var canvas: Canvas? = null
     var longPress: LongClickHandler? = null
@@ -49,7 +48,11 @@ open class FlutterMarker(mapView: MapView) : Marker(mapView), Marker.OnMarkerCli
 
     }
 
-    constructor(application: Application, mapView: MapView, point: GeoPoint) : this(mapView = mapView) {
+    constructor(
+        application: Application,
+        mapView: MapView,
+        point: GeoPoint
+    ) : this(mapView = mapView) {
         this.application = application
         this.mapView = mapView
         this.mPosition = point
@@ -57,7 +60,13 @@ open class FlutterMarker(mapView: MapView) : Marker(mapView), Marker.OnMarkerCli
             onMarkerClick(marker, map)
         }
         createWindowInfoView()
-        setInfoWindow(FlutterInfoWindow(infoView = infoWindow!!, mapView = mapView, point = mPosition))
+        setInfoWindow(
+            FlutterInfoWindow(
+                infoView = infoWindow!!,
+                mapView = mapView,
+                point = mPosition
+            )
+        )
     }
 
     constructor(mapView: MapView, infoWindow: View) : this(mapView) {
@@ -71,45 +80,65 @@ open class FlutterMarker(mapView: MapView) : Marker(mapView), Marker.OnMarkerCli
     }
 
 
-
 //    override fun onLongPress(event: MotionEvent?, mapView: MapView?): Boolean {
 //        longPress?.let { it(this) }
 //        return super.onLongPress(event, mapView)
 //    }
 
-    fun setIconMaker(color: Int?, bitmap: Bitmap?) {
-        getDefaultIconDrawable(color, bitmap).also {
+    fun setIconMaker(color: Int?, bitmap: Bitmap?,angle: Double = 0.0) {
+        getDefaultIconDrawable(color, bitmap,angle).also {
             icon = it
         }
     }
 
     fun defaultInfoWindow() {
 
-        setInfoWindow(FlutterInfoWindow(infoView = infoWindow
-                ?: createWindowInfoView(), mapView = mapView, point = this.mPosition))
+        setInfoWindow(
+            FlutterInfoWindow(
+                infoView = infoWindow
+                    ?: createWindowInfoView(), mapView = mapView, point = this.mPosition
+            )
+        )
     }
 
 
-
-    private fun getDefaultIconDrawable(color: Int?, bitmap: Bitmap?): Drawable {
+    private fun getDefaultIconDrawable(
+        color: Int?,
+        bitmap: Bitmap?,
+        angle: Double = 0.0
+    ): Drawable {
         var iconDrawable: Drawable? = null
         bitmap?.let { b ->
-            iconDrawable = BitmapDrawable(mapView.resources, b)
+            iconDrawable = when (angle > 0.0) {
+                true -> BitmapDrawable(mapView.resources, rotateMarker(b, angle))
+                false -> BitmapDrawable(mapView.resources, b)
+            }
             iconDrawable = iconDrawable.apply {
                 color?.let { c ->
-                    this?.colorFilter = BlendModeColorFilterCompat.createBlendModeColorFilterCompat(c, BlendModeCompat.SRC_OVER)
+                    this?.colorFilter = BlendModeColorFilterCompat.createBlendModeColorFilterCompat(
+                        c,
+                        BlendModeCompat.SRC_OVER
+                    )
                 }
             }
 
         } ?: run {
-            iconDrawable = ContextCompat.getDrawable(application, R.drawable.ic_location_on_red_24dp)!!
+            iconDrawable =
+                ContextCompat.getDrawable(application, R.drawable.ic_location_on_red_24dp)!!
         }
         return iconDrawable!!
 
     }
 
+    private fun rotateMarker(bitmap: Bitmap, angle: Double): Bitmap {
+        val matrix = Matrix()
+        matrix.postRotate(angle.toFloat())
+        return Bitmap.createBitmap(bitmap, 0, 0, bitmap.width, bitmap.height, matrix, true)
+    }
+
     private fun createWindowInfoView(): View {
-        val inflater = application.getSystemService(Context.LAYOUT_INFLATER_SERVICE) as LayoutInflater
+        val inflater =
+            application.getSystemService(Context.LAYOUT_INFLATER_SERVICE) as LayoutInflater
         infoWindow = inflater.inflate(R.layout.infowindow, null)
         return infoWindow!!
     }