+++ lib/src/channel/osm_method_channel.dart
@@ -371,7 +371,7 @@ class MethodChannelOSM extends OSMPlatform {
     try {
       List<Map<String, double>> listGeos = [];
       for (GeoPoint p in pList) {
-        listGeos.add({"lon": p.longitude, "lat": p.latitude});
+        listGeos.add(p.toMap());
       }
       await _channels[idOSM]!
           .invokeMethod("staticPosition", {"id": id, "point": listGeos});